requirements.txt
astroid==2.0.4
atomicwrites==1.3.0
attrs==18.2.0
colorama==0.4.1
Django==2.1.2
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==6.0.0
pluggy==0.9.0
psycopg2-binary==2.7.5
py==1.8.0
pycodestyle==2.4.0
pytest==4.3.0
pytest-django==3.4.7
pytz==2018.5
six==1.11.0
whitenoise==4.1
wrapt==1.10.11
